[APP-0002] Add routing
