+++ app/controllers/base_controller.rb
@@ -1,15 +1,17 @@
 module Controllers
   class BaseController
-    attr_reader :request
+    attr_reader :request, :routes
 
     def initialize(request)
       @request = request
     end
 
-    private
-
     def build_response(status: 200, body:[])
       [status, { "Content-Type" => "text/html" }, body]
+    end
+
+    def not_found(status: 404)
+      [status, { "Content-Type" => "text/html" }, ["Not found"]]
     end
 
     def params
